PHP 8 working with strict_types = 1

+++ src/SearchIndex.php
@@ -617,7 +617,7 @@ class SearchIndex
 
             $retry++;
             $factor = ceil($retry / 10);
-            usleep($factor * $time); // 0.1 second
+            usleep((int) ($factor * $time)); // 0.1 second
         } while (true);
     }
